splitting utils out
