+++ src/companion.go
@@ -232,113 +232,3 @@ func (c *Companion) RunDiscordBot() error {
     select {}
 }
 
-func (companion *Companion) AmIPrimary(m *discordgo.MessageCreate) bool {
-    sendPrimary := true
-    if RoomPrimaries[m.ChannelID] != companion.CompanionId {
-        // We're not the primary
-        VerboseLog("%v is not primary for %v - %v is", companion.CompanionId, m.ChannelID, RoomPrimaries[m.ChannelID])
-        sendPrimary = false
-    } else {
-        VerboseLog("%v is primary for %v", companion.CompanionId, m.ChannelID)
-    }
-    return sendPrimary
-}
-
-func (c *Companion) GetRoomMembers(roomId string) []string {
-    roomInfo, err := c.NomiKin.RoomExists(&roomId)
-    if err != nil {
-        log.Printf("Error checking if room exists: %v\n", err)
-        return []string{}
-    }
-
-    if roomInfo == nil || roomInfo.Nomis == nil {
-        log.Printf("Room info or room members are nil for room %s\n", roomId)
-        return []string{}
-    }
-
-    var retMembers []string
-    for _, m := range roomInfo.Nomis {
-        if m.Uuid != "" {
-            retMembers = append(retMembers, m.Uuid)
-        }
-    }
-
-    VerboseLog("Room members for %v are: %v", roomId, retMembers)
-    return retMembers
-}
-
-func (c *Companion) CheckRoomStatus(roomId string) string {
-    endpoint := "https://api.nomi.ai/v1/rooms/" + roomId
-    repl, err := c.NomiKin.ApiCall(endpoint, "GET", nil)
-    if err != nil {
-        log.Printf("Error calling Nomi %v API: %v\n", c.CompanionId, err)
-    }
-
-    var resp map[string]interface{}
-    err = json.Unmarshal(repl, &resp)
-    if err != nil {
-        log.Printf("%v failed to decode JSON: %v\n", c.CompanionId, err)
-    }
-
-    status, ok := resp["status"].(string)
-    if !ok {
-        log.Printf("%v status field is missing\n", c.CompanionId)
-    }
-
-    return status
-}
-
-func (c *Companion) WaitForRoom(roomId string) bool {
-    waitFor := 45
-    waited := 0
-    for {
-        if waited > waitFor {
-            VerboseLog("companion.WaitForRoom took longer than 45 seconds - Nomi: %v - Room: %v", c.CompanionId, roomId)
-            return false
-        }
-        if c.CheckRoomStatus(roomId) == "Default" {
-            return true
-        }
-        time.Sleep(time.Second * 1)
-        waited++
-    }
-}
-
-func (companion *Companion) HandleMessageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
-    if m.Author.ID == companion.DiscordSession.State.User.ID {
-        // We don't have to send our companion their own messages
-        return
-    }
-
-    if companion.CompanionType == "NOMI" && companion.ChatStyle == "ROOMS" {
-        // If we're in Rooms mode, drop messages for which we don't have a room setup for
-        if companion.RoomObjects[m.ChannelID].Uuid == "" {
-            return
-        }
-        
-        // If the message is from a Nomi in the same room, we don't have to process it, because they already did
-        if m.Author.Bot {
-            sharedNomiRoom := false
-            roomMems := companion.GetRoomMembers(m.ChannelID)
-            for b, c := range Companions {
-                if b.State.User.ID == m.Author.ID && Contains(roomMems, c.CompanionId) {
-                    sharedNomiRoom = true
-                    VerboseLog("%v message from Companion %v is in the same room %v - not forwarding to the room", companion.CompanionId, c.CompanionId, m.ChannelID)
-                    break
-                }
-            }
-
-            if sharedNomiRoom {
-                return
-            }
-        }
-    }
-
-    message := QueuedMessage{
-        Message: m,
-        Companion: companion,
-    }
-
-    companion.Queue.Enqueue(message)
-}
-
